Refactor rest only classes
