+++ core/src/main/kotlin/builder/kord/RestOnlyBuilder.kt
@@ -1,0 +1,44 @@
+package dev.kord.core.builder.kord
+
+import dev.kord.common.entity.Snowflake
+import dev.kord.core.ClientResources
+import dev.kord.core.Kord
+import dev.kord.rest.ratelimit.ExclusionRequestRateLimiter
+import dev.kord.rest.request.KtorRequestHandler
+import dev.kord.rest.request.RequestHandler
+import io.ktor.client.*
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+
+public abstract class RestOnlyBuilder {
+    protected var handlerBuilder: (resources: ClientResources) -> RequestHandler =
+        { KtorRequestHandler(it.httpClient, ExclusionRequestRateLimiter(), token = it.token) }
+
+    /**
+     * The [CoroutineDispatcher] kord uses to launch suspending tasks. [Dispatchers.Default] by default.
+     */
+    public var defaultDispatcher: CoroutineDispatcher = Dispatchers.Default
+
+    /**
+     * The client used for building [Gateways][Gateway] and [RequestHandlers][RequestHandler]. A default implementation
+     * will be used when not set.
+     */
+    public var httpClient: HttpClient? = null
+
+    public var applicationId: Snowflake? = null
+
+    /**
+     * Configures the [RequestHandler] for the [RestClient].
+     *
+     * ```
+     * Kord(token) {
+     *     { resources -> KtorRequestHandler(resources.httpClient, ExclusionRequestRateLimiter()) }
+     * }
+     * ```
+     */
+    public fun requestHandler(handlerBuilder: (resources: ClientResources) -> RequestHandler) {
+        this.handlerBuilder = handlerBuilder
+    }
+
+    public abstract fun build(): Kord
+}
